Hacky fix for ref.set_target()
For a fastforward marge you must set the target of the branch to the
remote's target. However, when you use pygit2's implementation it
places the index into a weird state. My quick fix for this is to just
do a hard reset after updating the reference.

Attempt to find a better solution:
http://stackoverflow.com/questions/27749418/implementing-pull-with-pygit2